Adds some discontinuity checkers
This attempts to detect discontinuites in the potential, which can result from ions that rattle around inside the lattice too much, and thereby re-access stale lattice sites. This results generally in high-energy trajectories, which are unphysical.

This code detects those, and then marks them as a failed trajectory, with E=-500