More tweaks and documentation
Adds more comments to the code, describing what is going on.

Changed MAXDIV and MINDIV config options to DIST_SEARCH and FAILED_DE.

DIST_SEARCH is how far atoms are search for, approximately as a radius in A, note that this is also then inside the 5A grid cells.

FAILED_DE allows tweaking the condition for dropping a trajectory based on change in energy for the particle in a single step.

NWRITX -> SCAT_FLAG and NWRITY -> SCAT_TYPE
These were re-named to what they are currently being used for...