Modifications for consistant thermal runs
Modified things such that single shot runs can be done runs when temperature is not 0.

The thermalization seed is now dependant on ion and lattice site index, as well as the configured thermal seed.

As such, NITER was changed to Ion_Index, to allow specifying which ion is being re-run.

The level output in the data file was replaced with the ion index, and as such the python scripts have been updated to read this.