Tweaks to ion nearby checks
This now collects nearby sites, and then sorts them by distance, allowing checking less sites in hameq, but involving a sort operation.

This is done a sufficiently low number of times to not increase runtime, at least not for reasonably small numbers of runs.

This fixes most of the discontinuity issues, it seems an interaction distance of 10, and a max cell distance of 2 seems to remove the vast majority of such discontinuities