More command line argument work
Improvements to Command line argument parsing, added a class ArgValue to string_utils.h, this contains some functions for getting the value as double, bool, int and string, which can simplify checks elsewhere.

The default value for this is "", which corresponds to bool value of false.

Arguments for safari are now:
 -i [inputfile]
 -s - enables single shot mode
 -x/y [value] - Sets x/y start, if -s is present

Also removed some un-used functions from vec_math.h