Adds in a load-from-file option for potentials
this involved converting a bunch of things accessing sites and atoms to using the pointers instead.

the file format is lines of:
A B V F

where A and B are atomic symbols, V and F are potential and force values.
the ordering needs to be the same as the tables generated as the cache, but can be interlaced for different A/B combinations.

This should only contain A B, not B A pairs, as both use the same array, so if you include a line for Na Cu, do not also include lines for Cu Na

This allows generating arbitrary potentials externally, and then loading them in, so the code is no longer limited to specific potential types, this is needed for the cascade mode, where we need specific potentials for different parts of interaction